refactor: extract migration.py from __init__.py (#289)
Move config-entry version migration and legacy hybrid Modbus split into
a dedicated migration module. Re-export async_migrate_entry from the
package root for HA discovery and backward-compatible imports.

Completes the remaining extract tracked on #289 after heartbeat.py and
oauth_view.py (device_helpers.py was already in #297).
